Revert "OBMFULL-5394 Implement ehcache statistics API for the crash commands"
This reverts commit 478092e44e5c24e327e0be5f9a427e004e3cf13d.

+++ java/sync/opush/push-dao-ehcache/src/main/java/org/obm/push/store/ehcache/EhCacheStatistics.java
@@ -36,20 +36,20 @@ public interface EhCacheStatistics {
 	/**
 	 * @return disk gets count for a short time before this call
 	 */
-	long shortTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	int shortTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return disk gets count for a medium time before this call
 	 */
-	long mediumTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	int mediumTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return disk gets count for a long time before this call
 	 */
-	long longTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
+	int longTimeDiskGets(String storeName) throws StatisticsNotAvailableException;
 
 	/**
 	 * @return current heap memory took by the given store, in bytes
 	 */
-	long memorySizeInBytes(String storeName);
+	int memorySizeInBytes(String storeName);
 }
